Change percentage field's type to the decimal
