+++ HappyTravel.Hiroshima.Common/Models/Accommodations/Rooms/RoomPromotionalOffer.cs
@@ -15,7 +15,7 @@ namespace HappyTravel.Hiroshima.Common.Models.Accommodations.Rooms
         
         public DateTime ValidToDate { get; set; }
         
-        public double DiscountPercent { get; set; }
+        public decimal DiscountPercent { get; set; }
         
         public string? BookingCode { get; set; }
