+++ HappyTravel.Hiroshima.DirectManager/RequestValidators/PromotionalOfferValidator.cs
@@ -13,7 +13,7 @@ namespace HappyTravel.Hiroshima.DirectManager.RequestValidators
             RuleFor(offer => offer.ValidTo).NotEmpty();
             RuleFor(offer => offer.ValidFrom).LessThanOrEqualTo(offer => offer.ValidTo);
             RuleFor(offer => offer.BookByDate).LessThanOrEqualTo(offer => offer.ValidFrom);
-            RuleFor(offer => offer.DiscountPercent).GreaterThanOrEqualTo(0d).LessThanOrEqualTo(100d);
+            RuleFor(offer => offer.DiscountPercent).GreaterThanOrEqualTo(0m).LessThanOrEqualTo(100m);
             RuleFor(offer => offer.Description)
                 .AnyLanguage()
                 .When(offer => offer.Description != null);
